Agregada abstraccion EsDeAlumno
Cambio en Seguimiento
Cambio mapeo Seguimiento

+++ src/Secretaria.Domain/Escuela/Alumno.cs
@@ -6,9 +6,9 @@ namespace Secretaria.Domain.Escuela
 {
     public class Alumno
     {        
-        public int Libro { get; set; }
+        public short Libro { get; set; }
         
-        public int Folio { get; set; }        
+        public byte Folio { get; set; }        
 
         public Curso CursoActual { get; set; }
